Improve `smbclient` Logging Accuracy
Previously, the string representing the options being passed to `smbclient` were assembled separately between the logger and the actual call to invoke the command. This meant that the command shown in the logs wasn't the actual command being passed-in -- there were subtle differences. Now they're based off the same strings.